back up before fucking around with makefile

+++ ft_strncmp.c
@@ -6,15 +6,17 @@
 /*   By: hsilverb <hsilverb@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/10 15:52:12 by hsilverb          #+#    #+#             */
-/*   Updated: 2022/11/15 17:41:09 by hsilverb         ###   ########lyon.fr   */
+/*   Updated: 2022/11/18 17:18:06 by hsilverb         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stdio.h>
+#include <string.h>
 
 int	ft_strncmp(char *s1, char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while ((s1[i] || s2[i]) && i < n)
